Made folder structure more consistent. Extracted the cache into it's own decorator
