+++ app/Http/Services/QuoteFake.php
@@ -1,0 +1,9 @@
+<?php namespace App\Http\Services;
+
+class QuoteFake implements Quote
+{
+    public function getQuote(): \stdClass
+    {
+        return (object)['quote'=>'quote!'];
+    }
+}
